- Updating devDependencies - Renaming the Pastes and Components for kebab-case
