+++ package.json
@@ -10,13 +10,13 @@
     "test": "jest"
   },
   "dependencies": {
-    "@types/styled-components-react-native": "5.2.5",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.5",
     "styled-components": "6.1.14"
   },
   "devDependencies": {
+    "@types/styled-components-react-native": "5.2.5",
     "@babel/core": "7.25.2",
     "@babel/preset-env": "7.25.3",
     "@babel/runtime": "7.25.0",
